Use sub-routing for branch conflicts
