+++ ui/src/components/router/index.tsx
@@ -1,2 +1,3 @@
 export * from './withErrorBoundary';
 export * from './withSearchParamsValue';
+export * from './withParam';
